More refactoring of state and the remaining bits
